+++ packages/driver/src/cy/commands/navigation.ts
@@ -457,14 +457,14 @@ export default (Commands, Cypress, cy, state, config) => {
 
   Cypress.on('test:before:run', reset)
 
-  Cypress.on('stability:changed', (bool, event) => {
+  Cypress.on('stability:changed', async (bool, event) => {
     // only send up page loading events when we're
     // not stable!
-    stabilityChanged(Cypress, state, config, bool)
+    await stabilityChanged(Cypress, state, config, bool)
   })
 
-  Cypress.on('navigation:changed', (source, arg) => {
-    navigationChanged(Cypress, cy, state, source, arg)
+  Cypress.on('navigation:changed', async (source, arg) => {
+    await navigationChanged(Cypress, cy, state, source, arg)
   })
 
   Cypress.on('form:submitted', (e) => {
@@ -616,6 +616,10 @@ export default (Commands, Cypress, cy, state, config) => {
           cleanup()
         }
 
+        // Make sure the reload command can communicate with the AUT.
+        // if we failed for any other reason, we need to display the correct error to the user.
+        Cypress.ensure.commandCanCommunicateWithAUT(cy)
+
         return null
       })
     },
@@ -699,6 +703,9 @@ export default (Commands, Cypress, cy, state, config) => {
           if (typeof cleanup === 'function') {
             cleanup()
           }
+
+          // Make sure the go command can communicate with the AUT.
+          Cypress.ensure.commandCanCommunicateWithAUT(cy)
 
           return null
         })