+++ packages/extension/test/integration/v2/background_spec.js
@@ -1,3 +1,4 @@
+/* tslint:disable:no-empty */
 require('../../spec_helper')
 const _ = require('lodash')
 const http = require('http')