+++ packages/reporter/src/hooks/hook-model.ts
@@ -1,11 +1,10 @@
 import _ from 'lodash'
 import { observable, computed, makeObservable } from 'mobx'
 
-import { FileDetails } from '@packages/types'
-
-import { Alias } from '../instruments/instrument-model'
-import Err from '../errors/err-model'
-import CommandModel from '../commands/command-model'
+import type { FileDetails } from '@packages/types'
+import type { Alias } from '../instruments/instrument-model'
+import type Err from '../errors/err-model'
+import type CommandModel from '../commands/command-model'
 
 export type HookName = 'before all' | 'before each' | 'after all' | 'after each' | 'test body' | 'studio commands'