+++ packages/reporter/src/runnables/suite-model.ts
@@ -1,8 +1,9 @@
 import _ from 'lodash'
 import { computed, observable, makeObservable } from 'mobx'
 import Runnable, { RunnableProps } from './runnable-model'
-import TestModel, { TestProps } from '../test/test-model'
-import { TestState } from '@packages/types'
+import type TestModel from '../test/test-model'
+import type { TestProps } from '../test/test-model'
+import type { TestState } from '@packages/types'
 
 export interface SuiteProps extends RunnableProps {
   suites: Array<SuiteProps>