+++ packages/server/lib/socket-base.ts
@@ -164,9 +164,9 @@ export class SocketBase {
 
     automation.use({
       // @ts-ignore - this error is new, but not introduced in the most recent edit. TODO: fix
-      onPush: (message, data) => {
+      onPush: async (message, data) => {
         socketIo.emit('automation:push:message', message, data)
-        cdpIo.emit('automation:push:message', message, data)
+        await cdpIo.emit('automation:push:message', message, data)
       },
     })